Refactor the backEnd of the application to make it more scalable
